コレクションの作成方法を変更し、すべてのinteractionのハンドリングを効率化
