+++ src/events/interactions/StringSelectMenu.ts
@@ -1,0 +1,51 @@
+import { GetErrorChannel, GetLogChannel } from "@/lib/channelUtils";
+import { EmbedBuilder, MessageFlags, StringSelectMenuInteraction } from "discord.js";
+import config from "@/config";
+
+const StringSelectMenu = async (interaction: StringSelectMenuInteraction) => {
+  try {
+    const [prefix, id] = interaction.customId.split("_");
+    const executionDefine = interaction.client.stringSelectMenus.get(prefix);
+    if (!executionDefine) {
+      console.log(
+        `[${interaction.client.function.timeUtils.timeToJSTstamp(
+          Date.now(),
+          true
+        )} info] Not Found: ${interaction.customId}`
+      );
+      return;
+    }
+    await executionDefine.execute(interaction, id);
+  } catch (error) {
+    console.error(
+      `[${interaction.client.function.timeUtils.timeToJSTstamp(
+        Date.now(),
+        true
+      )} error]An Error Occured in ${interaction.customId}\nDetails:\n${error}`
+    );
+    const logEmbed = new EmbedBuilder()
+      .setTitle("ERROR - cmd")
+      .setDescription("```\n" + (error as any).toString() + "\n```")
+      .setColor(config.color.e)
+      .setTimestamp();
+
+    const channel = await GetErrorChannel(interaction.client);
+    if (channel) {
+      channel.send({ embeds: [logEmbed] });
+    }
+    const messageEmbed = new EmbedBuilder()
+      .setTitle("すみません。エラーが発生しました。")
+      .setDescription("```\n" + error + "\n```")
+      .setColor(config.color.e)
+      .setTimestamp();
+    if (interaction.channel && interaction.channel.isSendable()) {
+      await interaction.channel.send({ embeds: [messageEmbed] });
+    }
+    const logChannel = await GetLogChannel(interaction.client);
+    if (logChannel) {
+      logChannel.send({ embeds: [messageEmbed] });
+    }
+  }
+};
+
+export default StringSelectMenu;
