+++ src/selectMenus/types/SelectMenuDefineType.d.ts
@@ -1,0 +1,6 @@
+import { StringSelectMenuInteraction } from "discord.js";
+
+export interface StringSelectMenuDefineType {
+  name: string;
+  execute(interaction: StringSelectMenuInteraction, id: string): Promise<void>;
+}
