+++ src/types/client.d.ts
@@ -2,6 +2,7 @@ import { Collection } from "discord.js";
 import { Command } from "@/commands/types/Command";
 import timeUtils from "@/lib/timeUtils";
 import logUtils from "@/lib/dataUtils";
+import { StringSelectMenuDefineType } from "@/selectMenus/types/SelectMenuDefineType";
 
 declare module "discord.js" {
   interface Client {
@@ -12,5 +13,6 @@ declare module "discord.js" {
       timeUtils: typeof timeUtils;
       logUtils: typeof logUtils;
     };
+    stringSelectMenus: Collection<string, StringSelectMenuDefineType>;
   }
 }
